+++ src/app/data/db.service.ts
@@ -1,5 +1,5 @@
 import { Injectable, signal } from '@angular/core';
-import { Event, Day, Camp, Art, Pin, DataMethods, MapSet, GeoRef, Dataset, RSLEvent } from './models';
+import { Event, Day, Camp, Art, Pin, DataMethods, MapSet, GeoRef, Dataset, RSLEvent, TimeRange, GPSSet, MapPoint } from './models';
 import { call, registerWorker } from './worker-interface';
 import { daysUntil, noDate, now } from '../utils/utils';
 import { GpsCoord, Point } from '../map/geo.utils';
@@ -61,8 +61,8 @@ export class DbService {
     return until;
   }
 
-  public async findEvents(query: string, day: Date | undefined, category: string, coords: GpsCoord | undefined): Promise<Event[]> {
-    return await call(this.worker, DataMethods.FindEvents, query, day, category, coords);
+  public async findEvents(query: string, day: Date | undefined, category: string, coords: GpsCoord | undefined, timeRange: TimeRange | undefined): Promise<Event[]> {
+    return await call(this.worker, DataMethods.FindEvents, query, day, category, coords, timeRange);
   }
 
   public async getEventList(ids: string[]): Promise<Event[]> {
@@ -77,12 +77,21 @@ export class DbService {
     return await call(this.worker, DataMethods.GpsToPoint, gpsCoord);
   }
 
+  public async gpsToMapPoint(gpsCoord: GpsCoord, title: string | undefined): Promise<MapPoint> {
+    const point = await this.gpsToPoint(gpsCoord);
+    return { street: '', clock: '', x: point.x, y: point.y, gps: structuredClone(gpsCoord), info: (title) ? { title, location: '', subtitle: '' } : undefined }
+  }
+
   public async getPotties(): Promise<Pin[]> {
     return await call(this.worker, DataMethods.GetPotties);
   }
 
   public async getMapPoints(name: string): Promise<MapSet> {
     return await call(this.worker, DataMethods.GetMapPoints, name);
+  }
+
+  public async getGPSPoints(name: string, title: string): Promise<MapSet> {
+    return await call(this.worker, DataMethods.GetGPSPoints, name, title);
   }
 
   public async getRSL(terms: string, day: Date | undefined, gpsCoord: GpsCoord | undefined): Promise<RSLEvent[]> {
@@ -115,6 +124,14 @@ export class DbService {
 
   public async findEvent(uid: string): Promise<Event> {
     return await call(this.worker, DataMethods.FindEvent, uid);
+  }
+
+  public async getMapPointGPS(mapPoint: MapPoint): Promise<GpsCoord> {
+    return await call(this.worker, DataMethods.GetMapPointGPS, mapPoint);
+  }
+
+  public async setMapPointsGPS(mapPoints: MapPoint[]): Promise<MapPoint[]> {
+    return await call(this.worker, DataMethods.SetMapPointsGPS, mapPoints);
   }
 
   public async findCamp(uid: string): Promise<Camp> {