+++ src/app/geolocation/geo.service.ts
@@ -1,16 +1,18 @@
-import { Injectable } from '@angular/core';
+import { Injectable, signal } from '@angular/core';
 import { Geolocation } from '@capacitor/geolocation';
-import { GpsCoord, Point, gpsToMap, setReferencePoints } from '../map/geo.utils';
+import { GpsCoord, NoGPSCoord, Point } from '../map/geo.utils';
 import { DbService } from '../data/db.service';
-import { MapPoint } from '../data/models';
-import { getPoint, toClock, toStreetRadius } from '../map/map.utils';
 import { environment } from 'src/environments/environment';
 import { Capacitor } from '@capacitor/core';
+import { noDate, secondsBetween } from '../utils/utils';
 
 @Injectable({
   providedIn: 'root'
 })
 export class GeoService {
+  private gpsPosition = signal(NoGPSCoord());
+  private lastGpsUpdate: Date = noDate();
+  private hasPermission = false;
 
   constructor(private db: DbService) { }
 
@@ -34,27 +36,48 @@ export class GeoService {
 
   public async getPosition(): Promise<GpsCoord> {
     if (!Capacitor.isNativePlatform()) {
-      console.error(`On web we return the coord of center camp`)
-      return { lat: -119.21121456711064, lng: 40.780501492435846 };
+      console.error(`On web we return the coord of center camp`);
+      this.gpsPosition.set({ lng: -119.21121456711064, lat: 40.780501492435846 });
+      return this.gpsPosition();
     }
 
-    if (!await this.checkPermissions()) {
-      if (!await this.getPermission()) {
-        console.error(`User geolocation permission denied.`);
-        return { lat: 0, lng: 0 };
+    console.time('geo.permissions');
+    if (!this.hasPermission) {
+      if (!await this.checkPermissions()) {
+        if (!await this.getPermission()) {
+          console.error(`User geolocation permission denied.`);
+          this.gpsPosition.set(NoGPSCoord());
+          return NoGPSCoord();
+        }
       }
     }
+
+    console.timeEnd('geo.permissions');
     if (environment.gps) {
+      console.error(`Fake GPS position was returned ${environment.gps}`);
+      this.gpsPosition.set(environment.gps);
       return environment.gps; // Return a fake location
     }
 
+    if (secondsBetween(this.lastGpsUpdate, new Date()) < 10) {
+      return this.gpsPosition();
+    }
     const position = await Geolocation.getCurrentPosition({ enableHighAccuracy: true });
-    return { lat: position.coords.latitude, lng: position.coords.longitude };
+
+    let gps = { lat: position.coords.latitude, lng: position.coords.longitude };
+    if (environment.latitudeOffset && environment.longitudeOffset) {
+      const before = structuredClone(gps);
+      const after = { lat: gps.lat + environment.latitudeOffset, lng: gps.lng + environment.longitudeOffset };
+      gps = after;
+      console.error(`GPS Position was modified ${JSON.stringify(before)} to ${JSON.stringify(after)}`);
+    }
+    this.lastGpsUpdate = new Date();
+    this.gpsPosition.set(gps);
+    this.hasPermission = true;
+    return gps;
   }
 
-  public async placeOnMap(coord: GpsCoord, circleRadius: number): Promise<Point> {
-    const point = await this.db.gpsToPoint(coord);
-
-    return point;
+  public async gpsToPoint(coord: GpsCoord, circleRadius: number): Promise<Point> {
+    return await this.db.gpsToPoint(coord);
   }
 }