+++ src/environments/environment.ts
@@ -1,19 +1,9 @@
-// This file can be replaced during build by using the `fileReplacements` array.
-// `ng build` replaces `environment.ts` with `environment.prod.ts`.
-// The list of file replacements can be found in `angular.json`.
 
 export const environment = {
   production: false,
   simulatedTime: new Date('Mon Aug 28 2023 09:18:00 GMT-0700'),
-  gps: { lat: -119.21121456711064, lng: 40.780501492435846 },
+  gps: { lng: -119.21121456711064, lat: 40.780501492435846 },  
+  latitudeOffset: undefined,  
+  longitudeOffset: undefined,
   overrideLocations: true
 };
-
-/*
- * For easier debugging in development mode, you can import the following file
- * to ignore zone related error stack frames such as `zone.run`, `zoneDelegate.invokeTask`.
- *
- * This import should be commented out in production mode because it will have a negative impact
- * on performance if an error is thrown.
- */
-// import 'zone.js/plugins/zone-error';  // Included with Angular CLI.