+++ src/environments/environment.vegas.ts
@@ -2,7 +2,7 @@ export const environment = {
   production: true,
   simulatedTime: undefined,
   gps: undefined,
-  overrideLocations: false,
-  latitudeOffset: undefined,  
-  longitudeOffset: undefined,
+  overrideLocations: false,  
+  latitudeOffset: 4.61608809015,
+  longitudeOffset: -3.8730106216
 };